feat: make privacy policy page dynamic with configurable sections
